refactor: split words with dashes, next.js routing
