+++ src/app/(frontend)/components/about-us/CardLayout.tsx
