+++ src/app/(frontend)/components/about-us/DesktopCard.tsx
