+++ src/app/(frontend)/components/about-us/Hero.tsx
