+++ src/app/(frontend)/components/our-people/FramedImage.tsx
