+++ src/app/(frontend)/components/our-people/Hero.tsx
