+++ src/app/(frontend)/components/our-people/Players.tsx
