+++ src/app/(frontend)/components/our-story/Hero.tsx
