+++ src/app/(frontend)/components/our-story/TimelineItem.tsx
